+++ weasel/src/main/kotlin/src/evalClosure.kt
@@ -1,7 +1,9 @@
-package closure
+package src
 
 import kotlinx.collections.immutable.PersistentMap
 import kotlinx.collections.immutable.persistentHashMapOf
+import src.parsing.Lexer
+import src.parsing.Parser
 
 typealias Env = PersistentMap<String, Value>
 
@@ -69,93 +71,41 @@ fun evalBinaryNumber(v1: Value, v2: Value, f: (Int, Int) -> Int): Value {
 
 fun testEval(expr: String) {
     try {
-        println(eval(persistentHashMapOf("fix" to z), Parser(Lexer(expr)).parseExpr()))
-    } catch (ex: Exception) {
+        println(
+            eval(
+                persistentHashMapOf()
+                ,
+                Parser(
+                    Lexer(
+                        expr
+                    )
+                )
+                    .
+                    parseExpr(
+                    )
+            )
+        )
+
+    }
+    catch
+        (
+        ex
+                     :
+                     Exception) {
         println("Failed to eval with: ${ex.message}")
     }
 
 }
 
-// sum(0) == 0
-// sum(n) == n + sum(n - 1)
-
-fun sum(n: Int): Int =
-    if (n == 0) {
-        0
-    } else {
-        n + sum(n - 1)
-    }
-
-val z = eval(
-    persistentHashMapOf(),
-    Parser(Lexer("""\f => (\x => f(\v => x x v)) (\x => f(\v => x x v))""")).parseExpr()
-)
-
 fun main() {
-//    val identity = Expr.Lambda("x", Expr.Var("x"))
-//    test(Expr.Number(10))
-//    test(Expr.Var("x"))
-//    test(identity)
-//    test(Expr.Application(identity, Expr.Number(20)))
-//    test(
-//        Expr.Application(
-//            Expr.Lambda("x", Expr.Lambda("y", Expr.Var("x"))),
-//            Expr.Number(20)
-//        )
-//    )
-//    test(
-//        Expr.Application(
-//            Expr.Application(
-//                Expr.Lambda("x", Expr.Lambda("y", Expr.Var("x"))), Expr.Number(20)
-//            ), Expr.Number(0)
-//        )
-//    )
-//    test(Expr.Application(Expr.Number(5), Expr.Number(20)))
-//    test(Expr.Binary(Operator.Plus, Expr.Number(5), Expr.Number(10)))
-//    test(Expr.Binary(Operator.Plus, identity, Expr.Number(10)))
-//    test(Expr.If(Expr.Boolean(true), Expr.Number(42), Expr.Number(10)))
-//    test(Expr.If(Expr.Boolean(false), Expr.Number(42), Expr.Number(10)))
-//    test(Expr.Binary(Operator.Equals, Expr.Number(10), Expr.Number(10)))
-//    test(Expr.Binary(Operator.Equals, Expr.Number(10), Expr.Number(11)))
-//    test(
-//        Expr.If(
-//            Expr.Binary(Operator.Equals, Expr.Number(10), Expr.Number(10)),
-//            Expr.Number(42),
-//            Expr.Number(10)
-//        )
-//    )
-//
-//    println(sum(5)) // 1 + 2 + 3 + 4 + 5 == 15
-//
-//    val z = Parser(Lexer("""\f => (\x => f(\v => x x v)) (\x => f(\v => x x v))""")).parseExpr()
-//    val sumSingle = Parser(Lexer("""\f => \n => if n == 0 then 0 else n + f (n - 1)""")).parseExpr()
-//
-//    val sumLambda = Expr.Application(z, sumSingle)
-//
-//    test(Expr.Application(sumLambda, Expr.Number(5)))
-    testEval(
-        """
-        let add3 = \x => x + 3 in
-        let twice = \f => \x => f (f x) in
-        twice add3 10
-    """.trimIndent()
-    )
-
-    testEval(
-        """
-        let fib = fix \fib => \n => 
-          if n == 0 then 
-            1 
-          else if n == 1 then 
-            1 
-          else 
-            fib (n - 1) + fib (n - 2) in
-        fib 5
-        """.trimIndent()
-    )
-
-    // Homework: Fibonacci
-    // fib(0) = 1
-    // fib(1) = 1
-    // fib(n) = fib(n + -1) + fib(n + -2)
+"""
+    fun myFunction(input: String): JSONObject{
+        return input+=" ist cool"
+    }
+    let x = "Basti"
+    {
+        foo: "bar"
+        yeet: myFunction(x)
+    }
+""".trimIndent()
 }
